Update README.md
I brought in a pie chart to highlight two years, 1864 & 1865 and the top three causes of death for those years. These years are outstanding on my other charts and I brought in a slicer, you can pick a cause and it will expand that information on the bar chart and show the count, I created relationships in the model view, that was a lot of fun! I manipulated the raw data using the power editor query and youtube videos I've watched on intro to Power BI plus the first chapter of Power BI on DC. This was a lot of fun to work on, I really love the opportunity there is with creating visuals with this tool. Looking forward to using this more.